requirements.txt
Flask
Flask-Cors
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
pandas
pypdf
pytest